# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.CreateReceiptRequest:
    properties:
      amount:
        type: integer
      booking_id:
        type: integer
      payment_method:
        enum:
        - Credit
        - Debit
        - Bank
        - Transfer
        type: string
    required:
    - amount
    - booking_id
    - payment_method
    type: object
...
  handlers.GetReceiptsByDateRangeRequest:
    properties:
      end_date:
        type: string
      start_date:
        type: string
    required:
    - end_date
    - start_date
    type: object
...
  handlers.ProcessRefundRequest:
    properties:
      booking_id:
        type: integer
    required:
    - booking_id
    type: object
...
  models.Receipt:
    properties:
      amount:
        type: integer
      booking_id:
        type: integer
      issue_date:
        type: string
      payment_date:
        type: string
      payment_method:
        enum:
        - Credit
        - Debit
        - Bank
        - Transfer
        type: string
      receipt_id:
        type: integer
    required:
    - amount
    - payment_method
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for hotel booking system with transaction management and concurrency
    control
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Hotel Booking System API
  version: "1.0"
paths:
...
  /receipts:
    get:
      description: Get all receipts with pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all receipts
      tags:
      - receipts
    post:
      consumes:
      - application/json
      description: Process payment for a booking
      parameters:
      - description: Receipt details
        in: body
        name: receipt
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateReceiptRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Receipt'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a payment receipt
      tags:
      - receipts
  /receipts/{id}:
    get:
      description: Get a receipt by ID
      parameters:
      - description: Receipt ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Receipt'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a receipt
      tags:
      - receipts
  /receipts/booking/{bookingId}:
    get:
      description: Get a receipt associated with a booking
      parameters:
      - description: Booking ID
        in: path
        name: bookingId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Receipt'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get receipt by booking
      tags:
      - receipts
  /receipts/by-date:
    post:
      consumes:
      - application/json
      description: Get all receipts within a date range
      parameters:
      - description: Date range
        in: body
        name: dates
        required: true
        schema:
          $ref: '#/definitions/handlers.GetReceiptsByDateRangeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Receipt'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get receipts by date range
      tags:
      - receipts
  /receipts/refund:
    post:
      consumes:
      - application/json
      description: Process a refund for a booking
      parameters:
      - description: Refund details
        in: body
        name: refund
        required: true
        schema:
          $ref: '#/definitions/handlers.ProcessRefundRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Process a refund
      tags:
      - receipts
...
swagger: "2.0"